sererate os and kernel stuff
